refactor: 엔티티명 수정 및 command / query 패키지 분리
