+++ src/main/java/com/example/hotdeal/domain/product/product/domain/command/Product.java
@@ -1,4 +1,4 @@
-package com.example.hotdeal.domain.product.product.domain;
+package com.example.hotdeal.domain.product.product.domain.command;
 
 import com.example.hotdeal.global.model.BaseEntity;
 import jakarta.persistence.*;
