+++ src/main/java/com/example/hotdeal/domain/product/product/domain/command/ProductStatistics.java
@@ -1,4 +1,4 @@
-package com.example.hotdeal.domain.product.product.domain;
+package com.example.hotdeal.domain.product.product.domain.command;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
